+++ spark-bigquery-connector-common/src/test/java/com/google/cloud/spark/bigquery/integration/WriteIntegrationTestBase.java
@@ -19,6 +19,7 @@ import static com.google.common.truth.Truth.assertThat;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeThat;
 
 import com.google.cloud.bigquery.BigQuery;
@@ -27,6 +28,7 @@ import com.google.cloud.bigquery.Field;
 import com.google.cloud.bigquery.Field.Mode;
 import com.google.cloud.bigquery.LegacySQLTypeName;
 import com.google.cloud.bigquery.QueryJobConfiguration;
+import com.google.cloud.bigquery.RangePartitioning;
 import com.google.cloud.bigquery.Schema;
 import com.google.cloud.bigquery.StandardTableDefinition;
 import com.google.cloud.bigquery.Table;
@@ -41,6 +43,7 @@ import com.google.cloud.spark.bigquery.integration.model.Data;
 import com.google.cloud.spark.bigquery.integration.model.Friend;
 import com.google.cloud.spark.bigquery.integration.model.Link;
 import com.google.cloud.spark.bigquery.integration.model.Person;
+import com.google.cloud.spark.bigquery.integration.model.RangeData;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.ProvisionException;
@@ -1050,6 +1053,41 @@ abstract class WriteIntegrationTestBase extends SparkBigQueryIntegrationTestBase
 
     Dataset<Row> readDF = spark.read().format("bigquery").load(table);
     assertThat(readDF.count()).isEqualTo(3);
+  }
+
+  @Test
+  public void testPartitionRange() {
+    // partition write not supported in BQ Storage Write API
+    assumeThat(writeMethod, equalTo(SparkBigQueryConfig.WriteMethod.INDIRECT));
+
+    List<RangeData> data =
+        Arrays.asList(new RangeData("a", 1L), new RangeData("b", 5L), new RangeData("c", 11L));
+    Dataset<Row> df = spark.createDataset(data, Encoders.bean(RangeData.class)).toDF();
+    String table = testDataset.toString() + "." + testTable + "_range";
+    df.write()
+        .format("bigquery")
+        .option("temporaryGcsBucket", TestConstants.TEMPORARY_GCS_BUCKET)
+        .option("partitionField", "rng")
+        .option("partitionRangeStart", "1")
+        .option("partitionRangeEnd", "21")
+        .option("partitionRangeInterval", "2")
+        .option("partitionRequireFilter", "true")
+        .option("table", table)
+        .option("writeMethod", writeMethod.toString())
+        .save();
+
+    Dataset<Row> readDF = spark.read().format("bigquery").load(table);
+    assertThat(readDF.count()).isEqualTo(3);
+    Table bqTable = bq.getTable(TableId.of(testDataset.toString(), testTable + "_range"));
+    assertThat(bqTable).isNotNull();
+    assertTrue(bqTable.getDefinition() instanceof StandardTableDefinition);
+    StandardTableDefinition bqTableDef = bqTable.getDefinition();
+    assertThat(bqTableDef.getRangePartitioning()).isNotNull();
+    RangePartitioning.Range expectedRange =
+        RangePartitioning.Range.newBuilder().setStart(1L).setEnd(21L).setInterval(2L).build();
+    String expectedField = "rng";
+    assertThat(bqTableDef.getRangePartitioning().getRange()).isEqualTo(expectedRange);
+    assertThat(bqTableDef.getRangePartitioning().getField()).isEqualTo(expectedField);
   }
 
   @Test